+++ src/shape_sensitivity/descriptors.h
@@ -7,7 +7,7 @@
 #include "art.cuh"
 #include "gpu/gpu_image.cuh"
 
-std::vector<double> calculateIARTD(img_desc* img_desc_gpu,
-                                   gpu_cost_function::GPUImage* dev_image);
+std::vector<float> calculateIARTD(img_desc* img_desc_gpu,
+                                  gpu_cost_function::GPUImage* dev_image);
 
 #endif  // DESCRIPTORS_H_
